more output-related cleanup
